backend/requirements.txt
Django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
channels==4.3.1
whitenoise==6.6.0
python-decouple==3.8
Pillow==11.0.0
pytest==9.0.1
pytest-django==4.11.1
Faker==38.0.0
freezegun==1.5.5
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pymysql==1.1.0
gunicorn==23.0.0
daphne==4.0.0
<<<<<<< HEAD
django-allauth==0.62.1
=======
openpyxl==3.1.5
reportlab==4.2.5
>>>>>>> 13a4e22ac92d7824c227a4dff1ae74d9d5e9cb09
